short->int

+++ src/compute_density.cpp
@@ -198,13 +198,13 @@ double integrate_triangle_left_one_cell(const dp_t &bv, const dp_t &uv, double h
 }
 
 double integrate_chanel_slant_right(const dp_t& bv, const dp_t& uv,
-        short curr_i, short next_i, const ip_t &sx, double b, const ip_t &sb,
+        int curr_i, int next_i, const ip_t &sx, double b, const ip_t &sb,
         const ip_t &sy) {
     if (fabs(uv.y - bv.y) <= FLT_MIN) return fabs(uv.y - bv.y);
 
     double result = 0, gx = 0, hx = 0;
     dp_t mv, rv;
-    short m_i = 0;
+    int m_i = 0;
     if (uv.x <= bv.x) {
         mv = uv;
         rv = bv;
@@ -249,12 +249,12 @@ double integrate_chanel_slant_right(const dp_t& bv, const dp_t& uv,
 // BOTTOMLEFTTR
 
 double integrate_chanel_slant_left(const dp_t& bv, const dp_t& uv,
-        short curr_i, short next_i, const ip_t &sx, const ip_t &sy,
+        int curr_i, int next_i, const ip_t &sx, const ip_t &sy,
         double b, const ip_t &sb) {
     if (fabs(uv.y - bv.y) <= FLT_MIN) return fabs(uv.y - bv.y);
 
     dp_t lv, mv; //   -  Left and middle vertices.
-    short m_i = 0; //   -  Where middle vertex is.        
+    int m_i = 0; //   -  Where middle vertex is.        
     double result = 0, gx = 0, hx = 0; //   -  Left and right boundary for each integration.   
 
     // зачем то определили среднюю точку
@@ -302,17 +302,17 @@ double integrate_right_triangle_bottom_left(const dp_t& bv, const dp_t& uv) {
 
     //   -  Index of current square by Ox and Oy axes. 
     ip_t sx, sy;
-    sx.x = static_cast<short> ((bv.x - FLT_MIN) / HX);
+    sx.x = static_cast<int> ((bv.x - FLT_MIN) / HX);
     if (bv.x - FLT_MIN <= 0) sx.x -= 1;
     sx.y = sx.x + 1;
-    sy.x = static_cast<short> ((bv.y + FLT_MIN) / HY);
+    sy.x = static_cast<int> ((bv.y + FLT_MIN) / HY);
     if (bv.y + FLT_MIN <= 0) sy.x -= 1;
     sy.y = sy.x + 1;
 
     ip_t ib(sx.x, sx.x + 1); //   -  Index of right boundary.   
 
     double result = 0;
-    short curr_i = 0, next_i = 0;
+    int curr_i = 0, next_i = 0;
     dp_t curr = bv, next;
     while (true) {
         //TODO: sx.x и sx.y должны быть положительными всегда? Кажется для sx.x это всегда верно...
@@ -355,16 +355,16 @@ double integrate_right_triangle_bottom_right(const dp_t& bv, const dp_t& uv) {
     if (!try_get_slope_ratio(bv, uv, k)) return k;
 
     ip_t sx, sy;
-    sx.x = static_cast<short> ((bv.x + FLT_MIN) / HX);
+    sx.x = static_cast<int> ((bv.x + FLT_MIN) / HX);
     if (bv.x + FLT_MIN <= 0) sx.x -= 1;
     sx.y = sx.x + 1;
-    sy.x = static_cast<short> ((bv.y + FLT_MIN) / HY);
+    sy.x = static_cast<int> ((bv.y + FLT_MIN) / HY);
     if (bv.y + FLT_MIN <= 0) sy.x -= 1;
     sy.y = sy.x + 1;
 
     ip_t ib(sx.x, sx.x + 1);
     double result = 0;
-    short curr_i = 0, next_i = 0;
+    int curr_i = 0, next_i = 0;
     dp_t curr = bv, next;
     while (true) {
         double slope = sy.y >= 0 ? fabs(OY[sy.y] - curr.y) : fabs(HY * sy.y - curr.y);
@@ -404,18 +404,18 @@ double integrate_right_triangle_upper_left(const dp_t& bv, const dp_t& uv) {
     if (!try_get_slope_ratio(bv, uv, k)) return k;
 
     ip_t sx, sy, ib;
-    sx.x = static_cast<short> ((bv.x + FLT_MIN) / HX); //   -  If bv.x is in grid edge I want it will be in the right side.
+    sx.x = static_cast<int> ((bv.x + FLT_MIN) / HX); //   -  If bv.x is in grid edge I want it will be in the right side.
     if (bv.x + FLT_MIN <= 0) sx.x -= 1;
     sx.y = sx.x + 1;
-    sy.x = static_cast<short> ((bv.y + FLT_MIN) / HY); //   -  If bv.y is in grid edge I want it will be in the upper square.
+    sy.x = static_cast<int> ((bv.y + FLT_MIN) / HY); //   -  If bv.y is in grid edge I want it will be in the upper square.
     if (bv.y + FLT_MIN <= 0) sy.x -= 1;
     sy.y = sy.x + 1;
-    ib.x = static_cast<short> ((uv.x - FLT_MIN) / HY); //   -  If uv.x is in grid edge I want it will be in the left side.
+    ib.x = static_cast<int> ((uv.x - FLT_MIN) / HY); //   -  If uv.x is in grid edge I want it will be in the left side.
     if (uv.x - FLT_MIN <= 0) ib.x -= 1;
     ib.y = ib.x + 1;
 
     double result = 0;
-    short curr_i = 0, next_i = 0;
+    int curr_i = 0, next_i = 0;
     dp_t curr = bv, next;
     while (true) {
         double slope = sy.y >= 0 ? OY[sy.y] - curr.y : fabs(HY * sy.y - curr.y);
@@ -456,18 +456,18 @@ double integrate_right_triangle_upper_right(const dp_t& bv, const dp_t& uv) {
     if (!try_get_slope_ratio(bv, uv, k)) return k;
 
     ip_t sx, sy, ib;
-    sx.x = static_cast<short> ((bv.x - FLT_MIN) / HX); //   -  If bv.x is in grid edge I want it will be between in the left side.
+    sx.x = static_cast<int> ((bv.x - FLT_MIN) / HX); //   -  If bv.x is in grid edge I want it will be between in the left side.
     if (bv.x - FLT_MIN <= 0) sx.x -= 1;
     sx.y = sx.x + 1;
-    sy.x = static_cast<short> ((bv.y + FLT_MIN) / HY); //   -  If bv.y is in grid edge I want it will be in the upper side.
+    sy.x = static_cast<int> ((bv.y + FLT_MIN) / HY); //   -  If bv.y is in grid edge I want it will be in the upper side.
     if (bv.y + FLT_MIN <= 0) sy.x -= 1;
     sy.y = sy.x + 1;
-    ib.x = static_cast<short> ((uv.x + FLT_MIN) / HX);
+    ib.x = static_cast<int> ((uv.x + FLT_MIN) / HX);
     if (uv.x + FLT_MIN <= 0) ib.x -= 1;
     ib.y = ib.x + 1;
 
     double result = 0;
-    short curr_i = 0, next_i = 0;
+    int curr_i = 0, next_i = 0;
     dp_t curr = bv, next;
     while (true) {
         double slope = sy.y >= 0 ? fabs(OY[sy.y] - curr.y) : fabs(HY * sy.y - curr.y);
